Split instances between core and data. Free now only depends on core.
