+++ modules/core/arrow-data/build.gradle
@@ -8,7 +8,7 @@ dependencies {
     testCompileOnly project(':arrow-annotations-processor')
     testCompile "io.kotlintest:kotlintest:$kotlinTestVersion"
     testCompile project(':arrow-syntax')
-    testCompile project(':arrow-instances')
+    testCompile project(':arrow-instances-data')
     testCompile project(':arrow-test')
     testCompile project(':arrow-free')
     testCompile project(':arrow-mtl')
