+++ modules/core/arrow-instances-core/build.gradle
@@ -2,7 +2,7 @@ dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     compile project(':arrow-annotations')
     compile project(':arrow-typeclasses')
-    compile project(':arrow-data')
+    compile project(':arrow-core')
     kapt project(':arrow-annotations-processor')
     kaptTest project(':arrow-annotations-processor')
     compileOnly project(':arrow-annotations-processor')
