+++ modules/core/arrow-instances-core/gradle.properties
