+++ modules/core/arrow-instances-core/src/main/kotlin/arrow/instances/function0.kt
@@ -1,8 +1,8 @@
 package arrow.instances
 
-import arrow.*
+import arrow.Kind
 import arrow.core.*
-import arrow.data.*
+import arrow.instance
 import arrow.typeclasses.*
 
 @instance(Function0::class)
