+++ modules/core/arrow-instances-core/src/main/kotlin/arrow/instances/number.kt
