+++ modules/core/arrow-instances-data/build.gradle
@@ -1,7 +1,9 @@
 dependencies {
-    compile project(':arrow-instances-data')
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     compile project(':arrow-annotations')
+    compile project(':arrow-typeclasses')
+    compile project(':arrow-data')
+    compile project(':arrow-instances-core')
     kapt project(':arrow-annotations-processor')
     kaptTest project(':arrow-annotations-processor')
     compileOnly project(':arrow-annotations-processor')
@@ -10,7 +12,6 @@ dependencies {
     testCompile project(':arrow-test')
 }
 
-// FIXME(paco): re-enable when it's prod-ready
-//apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
+apply from: rootProject.file('gradle/gradle-mvn-push.gradle')
 apply from: rootProject.file('gradle/generated-kotlin-sources.gradle')
 apply plugin: 'kotlin-kapt'
