+++ modules/core/arrow-instances-data/gradle.properties
@@ -1,0 +1,4 @@
+# Maven publishing configuration
+POM_NAME=Arrow Type Classes Instances
+POM_ARTIFACT_ID=arrow-instances
+POM_PACKAGING=jar
