+++ modules/core/arrow-instances-data/src/main/kotlin/arrow/instances/eithert.kt
