+++ modules/core/arrow-mtl/build.gradle
@@ -1,7 +1,7 @@
 dependencies {
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     compile project(':arrow-annotations')
-    compile project(':arrow-instances')
+    compile project(':arrow-instances-data')
     compile project(':arrow-free')
     kapt project(':arrow-annotations-processor')
     kaptTest project(':arrow-annotations-processor')
