+++ modules/docs/arrow-examples/src/test/kotlin/arrow/Typeclassless.kt
@@ -6,8 +6,6 @@ import arrow.TypeclasslessExamples.ScopeTwo.withApplicative
 import arrow.core.identity
 import arrow.data.ForListK
 import arrow.data.ListK
-import arrow.data.applicative
-import arrow.data.k
 import arrow.typeclasses.Applicative
 import io.kotlintest.matchers.shouldBe
 import io.kotlintest.specs.FreeSpec
