+++ modules/optics/arrow-optics/build.gradle
@@ -1,5 +1,7 @@
 dependencies {
     compile project(':arrow-free')
+    compile project(':arrow-data')
+    compile project(':arrow-instances-data')
     compile "org.jetbrains.kotlin:kotlin-stdlib-jdk7:$kotlinVersion"
     testCompile "io.kotlintest:kotlintest:$kotlinTestVersion"
     testCompile project(':arrow-test')
