+++ modules/optics/arrow-optics/src/main/kotlin/arrow/optics/instances/list_java.kt
@@ -21,7 +21,7 @@ interface ListEachInstance<A> : Each<List<A>, A> {
       ListK.traverse().run {
         FA.run {
           s.k().traverse(this, f).let {
-            it.map() {
+            it.map {
               it.list
             }
           }
@@ -43,7 +43,7 @@ interface ListFilterIndexInstance<A> : FilterIndex<List<A>, Int, A> {
             if (p(j)) f(a) else just(a)
           })
             .let {
-              it.map() {
+              it.map {
                 it.list
               }
             }
