+++ modules/optics/arrow-optics/src/test/kotlin/arrow/optics/GetterTest.kt
@@ -2,8 +2,6 @@ package arrow.optics
 
 import arrow.core.*
 import arrow.data.ListK
-import arrow.data.k
-import arrow.data.monoid
 import arrow.instances.StringMonoidInstance
 import arrow.test.UnitSpec
 import io.kotlintest.KTestJUnitRunner
