+++ settings.gradle
@@ -39,7 +39,8 @@ include {
             _ 'core'
             _ 'data'
             _ 'free'
-            _ 'instances'
+            _ 'instances-core'
+            _ 'instances-data'
             _ 'kindedj'
             _ 'mtl'
             _ 'syntax'
